requirements.txt
flask
pillow
pytest